+++ app/src/pages/experiment/ExperimentCompareMetricsPage.tsx
@@ -1,7 +1,10 @@
 import { css } from "@emotion/react";
 
 import { Flex, Heading, Icon, Icons, Text, View } from "@phoenix/components";
-import { costFormatter } from "@phoenix/utils/numberFormatUtils";
+import {
+  costFormatter,
+  latencyMsFormatter,
+} from "@phoenix/utils/numberFormatUtils";
 
 const metricCardCss = css`
   padding: var(--ac-global-dimension-size-200);
@@ -45,24 +48,27 @@ export function ExperimentCompareMetricsPage() {
           <div css={metricCardCss}>
             <Flex direction="column" gap="size-200">
               <Heading level={2}>Latency</Heading>
-              <BaseExperimentMetric value={520} formatter={formatLatencyMs} />
+              <BaseExperimentMetric
+                value={520}
+                formatter={latencyMsFormatter}
+              />
               <CompareExperimentMetric
                 value={620}
-                formatter={formatLatencyMs}
+                formatter={latencyMsFormatter}
                 baselineValue={520}
                 numImprovements={10}
                 numRegressions={5}
               />
               <CompareExperimentMetric
-                value={5.4 * MS_PER_SECOND}
-                formatter={formatLatencyMs}
+                value={5.4 * 1000}
+                formatter={latencyMsFormatter}
                 baselineValue={520}
                 numImprovements={0}
                 numRegressions={10}
               />
               <CompareExperimentMetric
-                value={1 * MS_PER_HOUR + 10 * MS_PER_SECOND}
-                formatter={formatLatencyMs}
+                value={1 * 60 * 60 * 1000 + 10 * 1000}
+                formatter={latencyMsFormatter}
                 baselineValue={520}
                 numImprovements={10}
                 numRegressions={0}
@@ -291,32 +297,4 @@ function ImprovementAndRegressionCounter({
       )}
     </Flex>
   );
-}
-
-const MS_PER_SECOND = 1000;
-const MS_PER_MINUTE = 60 * MS_PER_SECOND;
-const MS_PER_HOUR = 60 * MS_PER_MINUTE;
-
-/**
- * Formats a latency (given in milliseconds) to be displayed across different scales.
- * @param number
- * @returns {string} the string representation of the number
- */
-export function formatLatencyMs(number: number): string {
-  const hours = Math.floor(number / MS_PER_HOUR);
-  const minutes = Math.floor((number % MS_PER_HOUR) / MS_PER_MINUTE);
-  const seconds = Math.floor((number % MS_PER_MINUTE) / MS_PER_SECOND);
-  const milliseconds = Math.floor(number % MS_PER_SECOND);
-
-  if (hours > 0) {
-    return `${hours}h${minutes ? ` ${minutes}m` : ""}${seconds ? ` ${seconds}s` : ""}`;
-  }
-  if (minutes > 0) {
-    return `${minutes}m${seconds ? ` ${seconds}s` : ""}`;
-  }
-  if (seconds > 0) {
-    const tenthSeconds = Math.floor(milliseconds / 100);
-    return `${seconds}.${tenthSeconds.toFixed(0)}s`;
-  }
-  return `${milliseconds.toFixed(0)}ms`;
 }
